refactor: Move uploader package to internal
