+++ lib-gphotos/internal/uploader/errors.go
