+++ lib-gphotos/internal/uploader/progress.go
