+++ lib-gphotos/internal/uploader/progress_test.go
