+++ lib-gphotos/internal/uploader/requests.go
