+++ lib-gphotos/internal/uploader/requests_test.go
