+++ lib-gphotos/internal/uploader/store.go
