+++ lib-gphotos/internal/uploader/uploader.go
@@ -8,7 +8,7 @@ import (
 
 const (
 	// Prefix to write at beginning of each log line.
-	logPrefix = "gphotos-uploader:"
+	logPrefix = "gphotos-uploader: "
 
 	// Define which text to prefix to each log entry generated by the Logger.
 	// For example, flags Ldate | Ltime (or LstdFlags) produce,
