+++ lib-gphotos/internal/uploader/uploader_test.go
